3208: Removed refactorings from feature
