+++ src/Feed/CalendarApiFeedType.php
@@ -2,12 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Feed\SourceType\Calendar;
+namespace App\Feed;
 
 use App\Entity\Tenant\Feed;
 use App\Entity\Tenant\FeedSource;
-use App\Feed\FeedOutputModels;
-use App\Feed\FeedTypeInterface;
 use App\Feed\OutputModel\Calendar\Event;
 use App\Feed\OutputModel\Calendar\Location;
 use App\Feed\OutputModel\Calendar\Resource;
