+++ src/Feed/ColiboFeedType.php
@@ -2,13 +2,12 @@
 
 declare(strict_types=1);
 
-namespace App\Feed\SourceType\Colibo;
+namespace App\Feed;
 
 use App\Entity\Tenant\Feed;
 use App\Entity\Tenant\FeedSource;
-use App\Feed\FeedOutputModels;
-use App\Feed\FeedTypeInterface;
 use App\Feed\OutputModel\ConfigOption;
+use App\Feed\SourceType\Colibo\ApiClient;
 use App\Service\FeedService;
 use FeedIo\Feed\Item;
 use FeedIo\Feed\Node\Category;
