+++ src/Feed/KobaFeedType.php
@@ -2,12 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Feed\SourceType\Koba;
+namespace App\Feed;
 
 use App\Entity\Tenant\Feed;
 use App\Entity\Tenant\FeedSource;
-use App\Feed\FeedOutputModels;
-use App\Feed\FeedTypeInterface;
 use App\Service\FeedService;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpFoundation\Request;
