+++ src/Feed/SourceType/Colibo/ApiClient.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Feed\SourceType\Colibo;
 
 use App\Entity\Tenant\FeedSource;
+use App\Feed\ColiboFeedType;
 use Psr\Cache\CacheItemInterface;
 use Psr\Cache\CacheItemPoolInterface;
 use Psr\Log\LoggerInterface;
