+++ tests/Service/FeedServiceTest.php
@@ -6,13 +6,13 @@ namespace App\Tests\Service;
 
 use App\Entity\Tenant\Feed;
 use App\Entity\Tenant\FeedSource;
+use App\Feed\CalendarApiFeedType;
+use App\Feed\EventDatabaseApiFeedType;
 use App\Feed\FeedTypeInterface;
-use App\Feed\SourceType\Calendar\CalendarApiFeedType;
-use App\Feed\SourceType\EventDatabase\EventDatabaseApiFeedType;
-use App\Feed\SourceType\Koba\KobaFeedType;
-use App\Feed\SourceType\Notified\NotifiedFeedType;
-use App\Feed\SourceType\Rss\RssFeedType;
-use App\Feed\SourceType\SparkleIO\SparkleIOFeedType;
+use App\Feed\KobaFeedType;
+use App\Feed\NotifiedFeedType;
+use App\Feed\RssFeedType;
+use App\Feed\SparkleIOFeedType;
 use App\Service\FeedService;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
